refactor(analysis): extract cognitive complexity module
